Plumb through build client
