+++ src/Compilers/VisualBasic/VbcCore/VbcCore.csproj
@@ -59,6 +59,9 @@
     <Compile Include="..\..\Shared\BuildClient.cs">
       <Link>BuildClient.cs</Link>
     </Compile>
+    <Compile Include="..\..\Shared\CoreClrBuildHost.cs">
+      <Link>CoreClrBuildHost.cs</Link>
+    </Compile>
     <Compile Include="..\..\Shared\NativeMethods.cs">
       <Link>NativeMethods.cs</Link>
     </Compile>
