+++ src/Compilers/VisualBasic/vbc/vbc.csproj
@@ -60,6 +60,12 @@
     <Compile Include="..\..\Shared\BuildClient.cs">
       <Link>BuildClient.cs</Link>
     </Compile>
+    <Compile Include="..\..\Shared\DesktopBuildClient.cs">
+      <Link>DesktopBuildClient.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Shared\DesktopBuildHost.cs">
+      <Link>DesktopBuildHost.cs</Link>
+    </Compile>
     <Compile Include="..\..\Shared\NativeMethods.cs">
       <Link>NativeMethods.cs</Link>
     </Compile>
